Abstract out API call logic into service class
