+++ frontend/src/models/question.model.ts
@@ -1,0 +1,7 @@
+export interface Question {
+  id: number;
+  title: string;
+  description: string;
+  category: Array<string>;
+  complexity: string;
+}
