+++ frontend/src/pages/Home/Home.tsx
@@ -1,12 +1,11 @@
-import { ReactElement, useEffect, useState } from "react";
+import { ReactElement } from "react";
 //import { useNavigate } from "react-router-dom";
 import "./Home.scss";
 import Navbar from "../../components/Navbar/Navbar";
 import { Box, Button, Typography } from "@mui/material";
 import CodeEditorImage from "../../assets/code_editor.svg";
 import Footer from "../../components/Footer/Footer";
-import QuestionList, { Question } from "../../components/QuestionList/QuestionList";
-import axios from "axios";
+import QuestionList from "../../components/QuestionList/QuestionList";
 
 const Home = (): ReactElement => {
   // const navigate = useNavigate();
@@ -14,23 +13,6 @@ const Home = (): ReactElement => {
   // const goToQuestionList = () => {
   //   navigate("/questions");
   // };
-  const [questionList, setQuestionList] = useState<Question[]>([]);
-  const [loading, setLoading] = useState<boolean>(true);
-  const [error, setError] = useState<string | null>(null);
-
-  useEffect(() => {
-    // Fetch data from the backend REST API
-    axios
-      .get("/api/questions")
-      .then((response) => {
-        setQuestionList(response.data);
-        setLoading(false);
-      })
-      .catch((error) => {
-        setError("Failed to fetch questions");
-        setLoading(false);
-      });
-  }, []);
 
   return (
     <Box className="Home">
@@ -64,7 +46,7 @@ const Home = (): ReactElement => {
       </Box>
 
       <Box id="questions">
-        <QuestionList questionList={questionList} loading={loading} error={error} />
+        <QuestionList />
       </Box>
 
       {/* <button id="questions" onClick={goToQuestionList} className="Home-button">
